requirements.txt
asgiref==3.7.2
boto3==1.29.5
botocore==1.32.5
Django==4.2.5
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
jmespath==1.0.1
mysqlclient==2.1.1
packaging==23.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
whitenoise==6.6.0
